Bug 13881: Add desk management
Desks can be created for a branch. If a librarian sets a desk in the
intranet and check in a document on hold, the document is then
attached to a desk.

Test plan, basic desk creation/modification/deletion:

1/ perl installer/data/mysql/updatedatabase.pl (create new table desks)

2/ prove t/db_dependent/Desks.t

3/ go to admin page, click on desk management, try to create, modify
and delete desks

Test plan, setting desk in the intranet:

1/ Go to the intranet administration, remove all desks if you have
any. The desks should not appear in the header.

2/ add some desks in administration, without selecting some. The
message « NO DESK SET » should appear at the left of the branchname.

3/ set a desk for your session. The desk deskcode should appear beside
the branchname.

4/ change desk, the new name should appear also.

Test plan, attach holds awaiting pickup to a desk:

1/ Set a desk, put a book on hold, check it in, confirm hold, it
should appear in the "holds awaiting pickup" circulation page, it
should appear as waiting at the desk on the patron's session.

2/ Suppress desk. The hold is still pending, but desk mentions should
have disappear from "holds awaiting pickup" and patron's session.

http://bugs.koha-community.org/show_bug.cgi?id=13881

Fixed conflict in installer/data/mysql/kohastructure.sql
Followed testplan, works as advertised.
Signed-off-by: Marc Véron <veron@veron.ch>